requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
deep-translator==1.11.4
easyocr==1.7.2
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pluggy==1.6.0
pyclipper==1.3.0.post6
pytest==8.3.5
python-bidi==0.6.6
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.7.1
shapely==2.1.0
soupsieve==2.7
sympy==1.14.0
thefuzz==0.22.1
tifffile==2025.5.10
torch==2.7.0
torchvision==0.22.0
triton==3.3.0
typing_extensions==4.13.2
urllib3==2.4.0
